+++ src/artbox/__init__.py
@@ -9,7 +9,7 @@ def get_version():
     try:
         return importlib_metadata.version(__name__)
     except importlib_metadata.PackageNotFoundError:  # pragma: no cover
-        return "0.6.0"  # semantic-release
+        return "0.7.0"  # semantic-release
 
 
 version = get_version()